pyproject.toml
[project]
name = "glendza"
version = "0.1.0"
description = "My personal website."
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
  "Django==5.2.5",
  "django-environ>=0.11.2,<0.12",
  "django-icons>=25.2,<26",
  "wagtail==7.1",
  "wagtail-django-recaptcha>=2.1.1,<3",
  "whitenoise>=6.9.0,<7",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.16.0,<2.0.0",
  "ruff>=0.11.12,<1.0.0",
  "django-debug-toolbar>=5.2.0,<6",
]
prod = [
  "gunicorn==23.0.0",
]

[project.scripts]
gmanage = "glendza.web_app.manage:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/glendza"]
include = ["glendza/web_app/static/**"]

[tool.ruff]
line-length = 119

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true

[tool.djlint]
profile = "django"
ignore = "H006"
line_length = 119
indent = 2
indent_with_spaces = true
blank_line_after_tag = "load,extends,include"
blank_line_before_tag = "load,extends,include"
close_void_tags = true
format_js = true
format_css = true
format_attribute_template_tags = true
line_break_after_multiline_tag = true

[tool.djlint.js]
indent_size = 2

[tool.djlint.css]
indent_size = 2
